move LEB128 buffers to compatibility, flip defaults
